Move some components to the parent folder
