+++ src/renderer/components/PlaybackQueue.tsx
@@ -1,7 +1,7 @@
 import { DolphinEntry } from "@vinceau/slp-realtime";
 import React from "react";
 import { Droppable } from "react-beautiful-dnd";
-import { PlaybackQueueItem } from "./recorder/PlaybackQueueItem";
+import { PlaybackQueueItem } from "./PlaybackQueueItem";
 
 export const PlaybackQueue: React.FC<{
   id: string;
