+++ src/renderer/components/RecordButton.tsx
